+++ facetorch/analyzer/core.py
@@ -6,7 +6,6 @@ from codetiming import Timer
 from facetorch.analyzer.predictor.core import FacePredictor
 from facetorch.datastruct import ImageData, Response
 from facetorch.logger import LoggerJsonFile
-from facetorch.utils import draw_boxes_and_save
 from hydra.utils import instantiate
 from omegaconf import OmegaConf
 
@@ -16,13 +15,16 @@ logger = LoggerJsonFile().logger
 class FaceAnalyzer(object):
     @Timer("FaceAnalyzer.__init__", "{name}: {milliseconds:.2f} ms", logger.debug)
     def __init__(self, cfg: OmegaConf):
-        """FaceAnalyzer is the main class that reads images and runs face detection, tensor unification, as well as facial feature analysis prediction.
-        It is the orchestrator responsible for initializing and running the following components:
+        """FaceAnalyzer is the main class that reads images, runs face detection, tensor unification and facial feature prediction.
+        It also draws bounding boxes and facial landmarks over the image.
+
+        The following components are used:
 
         1. Reader - reads the image and returns an ImageData object containing the image tensor.
         2. Detector - wrapper around a neural network that detects faces.
         3. Unifier - processor that unifies sizes of all faces and normalizes them between 0 and 1.
-        4. Predictor list - list of wrappers around models trained to analyze facial features for example, expressions.
+        4. Predictor dict - dict of wrappers around neural networks trained to analyze facial features.
+        5. Utilizer dict - dict of utilizer processors that can for example extract 3D face landmarks or draw boxes over the image.
 
         Args:
             cfg (OmegaConf): Config object with image reader, face detector, unifier and predictor configurations.
@@ -33,7 +35,8 @@ class FaceAnalyzer(object):
             detector (FaceDetector): FaceDetector object that wraps a neural network that detects faces.
             unifier (FaceUnifier): FaceUnifier object that unifies sizes of all faces and normalizes them between 0 and 1.
             predictors (Dict[str, FacePredictor]): Dict of FacePredictor objects that predict facial features. Key is the name of the predictor.
-            logger (logging.Logger): Logger object that logs messages.
+            utilizers (Dict[str, FaceUtilizer]): Dict of FaceUtilizer objects that can extract 3D face landmarks, draw boxes over the image, etc. Key is the name of the utilizer.
+            logger (logging.Logger): Logger object that logs messages to the console or to a file.
 
         """
         self.cfg = cfg
@@ -51,12 +54,20 @@ class FaceAnalyzer(object):
         self.logger.info("Initializing FaceUnifier")
         self.unifier = instantiate(self.cfg.unifier)
 
-        self.logger.info("Initializing dictionary of FacePredictor objects")
+        self.logger.info("Initializing FacePredictor objects")
         self.predictors = {}
         for predictor_name in self.cfg.predictor:
             self.logger.info(f"Initializing FacePredictor {predictor_name}")
             self.predictors[predictor_name] = instantiate(
                 self.cfg.predictor[predictor_name]
+            )
+
+        self.logger.info("Initializing BaseUtilizer objects")
+        self.utilizers = {}
+        for utilizer_name in self.cfg.utilizer:
+            self.logger.info(f"Initializing BaseUtilizer {utilizer_name}")
+            self.utilizers[utilizer_name] = instantiate(
+                self.cfg.utilizer[utilizer_name]
             )
 
     @Timer("FaceAnalyzer.run", "{name}: {milliseconds:.2f} ms", logger.debug)
@@ -120,13 +131,18 @@ class FaceAnalyzer(object):
                 self.logger.info(f"Running FacePredictor: {predictor_name}")
                 data = _predict_batch(data, predictor, predictor_name)
 
-        path_output = None if path_output == "None" else path_output
-        if path_output is not None:
-            self.logger.info("Saving image", extra={"path_output": path_output})
-            draw_boxes_and_save(data, path_output)
+            path_output = None if path_output == "None" else path_output
+            data.path_output = path_output
+
+            self.logger.info("Utilizing facial features")
+            for utilizer_name, utilizer in self.utilizers.items():
+                self.logger.info(f"Running BaseUtilizer: {utilizer_name}")
+                data = utilizer.run(data)
 
         if not include_tensors:
-            self.logger.info("Removing tensors")
+            self.logger.debug(
+                "Removing tensors from response as include_tensors is False"
+            )
             data.reset_tensors()
 
         response = Response(faces=data.faces, version=data.version)