+++ tests/conftest.py
@@ -37,7 +37,19 @@ def pytest_configure(config):
         "markers", "fer: mark tests related to Facial Expression Recognition Predictor"
     )
     config.addinivalue_line(
-        "markers", "deepfake: mark tests related to Deepfake detection Predictor"
+        "markers", "deepfake: mark tests related to Deepfake Detection Predictor"
+    )
+    config.addinivalue_line(
+        "markers", "align: mark tests related to Face Alignment Predictor"
+    )
+    config.addinivalue_line(
+        "markers", "utilizer: mark tests related to the BaseUtilizer"
+    )
+    config.addinivalue_line(
+        "markers", "draw: mark tests related to the BoxDrawer utilizer"
+    )
+    config.addinivalue_line(
+        "markers", "save: mark tests related to the ImageSaver utilizer"
     )
     config.addinivalue_line(
         "markers", "downloader: mark tests related to the BaseDownloader"